requirements.txt
colorama==0.4.6
tomlkit==0.13.3
aiohttp==3.12.15
aiohttp_socks==0.10.1
requests==2.32.5
aiofiles==24.1.0
aiogram==3.22.0
discord-webhook==1.4.1
emoji==2.15.0
columnar==1.4.1
python-dotenv==1.0.0
